정말 삭제 할건지 confirm 추가

+++ src/components/TodoList.jsx
@@ -2,16 +2,20 @@ import { deleteDoc, doc } from "firebase/firestore";
 import { db } from "../firebase/config";
 
 export default function TodoList({ todos }) {
-  const handleClick = async (id) => {
-    const ref = doc(db, "todos", id);
-    await deleteDoc(ref);
+  const handleClick = async (id, title) => {
+    const shouldDelete = window.confirm(`"${title}" 항목을 삭제하시겠습니까?`);
+
+    if (shouldDelete) {
+      const ref = doc(db, "todos", id);
+      await deleteDoc(ref);
+    }
   };
 
   return (
     <div className="todo-list">
       <ul>
         {todos.map((todo) => (
-          <li key={todo.id} onClick={() => handleClick(todo.id)}>
+          <li key={todo.id} onClick={() => handleClick(todo.id, todo.title)}>
             {todo.title}
           </li>
         ))}
